Prevented consecutive moves from trying to move the same disc.
It is clearly not much good having consecutive moves attempt to move
the same disc. move now has two additional parameters, an LIn and
LOut. LIn is the last disc moved, and it is used to validate a
move. The disc moved is returned as LOut. This change improves the
solutions found.